crd-schema-gen: split code from main.go to make glide happy
